modified configs to work on Fedora 20

+++ install.sh
@@ -1,8 +1,8 @@
 [ -f ~/.bash_aliases ] && mv ~/.bash_aliases ~/.bash_aliases_old
 ln -s ~/repos/configs/conf-files/bash/.bash_aliases ~/.bash_aliases
 
-mv ~/.vimrc ~/.vimrc_old
+[ -f ~/.vimrc ] && mv ~/.vimrc ~/.vimrc_old
 ln -s ~/repos/configs/conf-files/.vimrc ~/.vimrc
 
-mv ~/.gitconfig ~/.gitconfig_old
+[ -f ~/.gitconfig ] && mv ~/.gitconfig ~/.gitconfig_old
 ln -s ~/repos/configs/conf-files/.gitconfig ~/.gitconfig
